Moved OperatorApi types to its own file to make exporting easier
